feat: 게임 로직 분리
